Testing/Testing-backend/backend copy/requirements.txt
flask==2.2.3
flask-cors==3.0.10
numpy==1.23.5
opencv-python-headless==4.7.0.72
pillow==9.5.0
tensorflow==2.12.0
joblib==1.2.0
gunicorn==20.1.0
Werkzeug==2.2.2
scikit-image==0.19.3